sipXportLib: Fix compilation warning.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@11407 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXportLib/src/os/OsSSLConnectionSocket.cpp
@@ -361,7 +361,7 @@ void OsSSLConnectionSocket::SSLInitSocket(int socket, long timeoutInSecs)
        {
           OsSysLog::add(FAC_KERNEL, PRI_DEBUG
                         , "OsSSLConnectionSocket bad parameters mSSL=%p,%d, closing socket..."
-                        , mSSL, socketDescriptor);
+                        , mSSL, (int)socketDescriptor);
           mIsConnected = FALSE;
           OsConnectionSocket::close();
           socketDescriptor = OS_INVALID_SOCKET_DESCRIPTOR;
